move output functions `free` -> `ops`
